pyproject.toml
[tool.poetry]
name = "data-axle-api-tool"
version = "0.1.0"
description = ""
authors = ["will-s-stone <will@williamstone.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.1"
requests = "^2.32.3"
python-dotenv = "^1.1.0"
shapely = "^2.1.0"
fastkml = "^1.1.0"
pyproj = "^3.7.1"
simplekml = "^1.3.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
